+++ src/app/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import './App.scss';
 
 function App(): React.ReactElement {
   return (
